Create certificate authority files in IMAGE_DATA directory

+++ ca/clean.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
 
-rm -rf certs crl intermediate newcerts private
-rm -f openssl.cnf index.txt{,.attr,.old} serial{,.old}
+
+if [ -d $IMAGE_DATA/ca ]
+then
+  cd $IMAGE_DATA/ca
+  rm -rf certs crl intermediate newcerts private
+  rm -f openssl.cnf index.txt{,.attr,.old} serial{,.old}
+fi
